pyproject.toml
[project]
name = "ai-specialization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.33.0",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "whoosh>=2.7.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
